requirements.txt
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
thefuzz[speedup]>=0.20.0
isodate>=0.6.1
rich>=13.0.0
